chore: 숫자 생성시, 숫자 이외의 값일 떄 예외 발생시키는 기능 추가

+++ src/main/kotlin/calculator/model/AddCalculator.kt
@@ -9,7 +9,7 @@ object AddCalculator {
         }
 
         return input.split(DEFAULT_DELIMITER_REGEX)
-            .map { Number.from(it.toInt()) }
+            .map { Number.from(it) }
             .sumOf { it.number }
     }
 }
